feat: add edit functionality
a functionality for editing the info has been implemented. the user can
now click on one of the section buttons to edit their information.

if the section to be editted has multiple indices of data stored, the
user is then prompted to specify the index they wish to edit. the index
of data is then rendered in the form inputs for the user to be editted.

this is a very sloppy feature as of right now, but the main idea is that
this is just an mvp. further refinements will follow Soon™