Refactor & Modularize the Code, Prevent Duplication (#1)
* refactor: complete refactor to modularize the code

completely refactoring the logic, states and utility functions away to
properly modularize the code. this is the first part of refactoring
where utils and `PersonalInfo` component was done.

the rest will follow.

* refactor: add `useForm` hooks & fix: `handleEdit`

the handleEdit logic was all over the place. this is no more. the edit
functionality is now working as intended. however, it will made more
streamlined for better ux.

* fix: remove the `useEffect` from `useForm`

the useEffect was only temporary and thus was removed as it is no longer
needed and was causing confusions.

* feat: add some error handlers & simplify code

* refactor: remove duplicate & clean up general code