feat: add support for adding new form entries
the form entries can now be appended into the form with the click of a
button. this required some refactors to some of the functions in utility
storage and `useForm` modules. this is only the bare implementation of
the feature, and further refinements such as `clearInputs` and other UX
improvements are underway.