feat: add support for multiple form entries
the user is now able to add multiple entries to the forms and have them
displayed in the `Resume.jsx` properly.

the user input is now saved inside an array instead of a plain object.
the `OutputSection.jsx` maps the data only if `Array.isArray` is true as
well as the array.lenght is greater than 0, conditionally render each
index.