feat: add `handleEdit` logic to access the data[i]
this feature currently only accesses the ith index of data that the user
wants to edit, prints it to the form inputs from state. however, this
function does not acutally edit the data as of yet. instead, it just
adds a brand new entry into the existing data array. the next commit is
hopefully going to resolve that.