refactor: `toggleOpenSections` to toggle on edit
the toggleOpenSections function now receives the `formSections.title`
instead of the index of the mapping as it previously were. the index was
local to the map and rendered it impossible for the function to be fired
from parent component(s). with this change, the `handleEditEntry` will
now be able to call it on edit.