refactor: split `OutputSection` to smaller chunks
the OutputSection was scraped and it was split into the following:

  - DataRow
  - ResumeAside
  - ResumeMain
  - SectionHeader
  - ContactInfo

all of which conditionally render the data the same way OutputSection
did, but much more cleanly and nicely.

also changed the layout of the resume to feature a main and aside to
display things a bit nicer. more inputs will be implemented to display
more things on the resume Soon™.