fix: implement merge logic for both states
both the old `personalInfo`, `educationInfo`, `workInfo` as well as the
newly introduced `liveFormData` states are now working in conjunction.
the user starts typing and their input will appear on the Résumé section
in real time; and once they submit their forms, the data will be saved,
not only to the sessionStorage, but also to their respecive afore
mentioned states, thus the data will always be displayed on the page
without any hiccups.

the edit and delete functions have not yet been tied to this new logic
so they will be the next.