feat: add misc form section with additional inputs
misc section contains 3 types of inputs, which have the functionality to
render a brand new one as the user types in and presses enter key. this
will continue generating additional inputs up to 5 inputs of each type,
for a total of 15 inputs.

the data for each type of inputs are added to their respective state
as array indices and are then stored as normal.

a pair of `<li>` elements are dynamically being rendered to display the
data in the aside section of the résumé.

further styling of these li elements will be implemented soon.

the current data structure of these states are a bit scuffed and look to
be prone to errors. this might or might not be fixed in the future, but
for the time being, it seems sufficient to do the job at minimum.