fix: various issues in data structures
this is getting a bit scuffect, the data structure and the way the state
is processed in one component is getting different from others. a lot of
data structure issues are taking place. this update attempts to fix many
of these issues for now, but who knows what will pop up in the future?